requirements.txt
pygame>=2.0
matplotlib>=3.0
pillow